test-requirements.txt
molecule
molecule-docker
docker
ansible-lint
yamllint
flake8
jmespath
# pytest==6.2.5
pytest-testinfra==6.4.0
tox
tox-gh-actions
